+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/user-settings/user-settings.tsx
@@ -16,7 +16,7 @@ import * as React from 'react'
 import ThemeSettings from '../../container/theme-settings'
 import AlertSettings from '../../alert-settings'
 import SearchSettings from '../../search-settings'
-import HiddenSettings from '../../../react-component/container/user-blacklist/user-blacklist'
+import HiddenSettings from '../../user-blacklist/container'
 const MapSettings = require('../../../component/layers/layers.view.js')
 import TimeSettings from '../../container/time-settings'
