+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/user-blacklist/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './user-blacklist'
+export { default } from './container'
