+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/user-blacklist/presentation.tsx
@@ -15,9 +15,9 @@
 
 import { hot } from 'react-hot-loader'
 import * as React from 'react'
-import styled, { keyframes, css } from 'styled-components'
-import { Button, buttonTypeEnum } from '../button'
-import BlacklistItemContainer from '../../blacklist-item'
+import styled, { keyframes } from 'styled-components'
+import { Button, buttonTypeEnum } from '../presentation/button'
+import BlacklistItemContainer from '../blacklist-item'
 
 const expandAnimation = keyframes`
   from {
